requirements.txt
pandas
matplotlib
tensorflow-gpu
keras
IPython
scikit-image
opencv-python
sklearn